refactor: collect dependencies
